refactor(backend): alinha o modelo de corrida ao artefato banco-e-integracoes
Segue a ordem de priorizar sempre o artefato e a versao mais completa.

- migracao 00008: runs recebe o conjunto completo de colunas de precisao
  do modelo (gap_pace_s, avg_hr/max_hr, calories, max_cadence_spm,
  stride_len_m, running_power_w, vert_osc_mm, ground_contact_ms,
  elev_loss_m, alt_min/max_m, best_km_pace_s, device_id); splits e
  best_efforts vao para run_tracks (splits_jsonb, best_efforts_jsonb),
  geom vira LineStringZ, +started_at (chave de particao futura);
  personal_records passa a distance_key/value_s (bigint); tabela
  run_metrics (introduzida em 00007, fora do modelo) e removida.
- metrics.go: novo bestEfforts() - melhor tempo continuo por distancia
  padrao (1k..42k) + maior distancia em 1h, para os recordes pessoais.
- store.go grava tudo nas colunas de runs + jsonb de run_tracks;
  GET /v1/runs/:id/metrics le de la.
- stats.go: personal_records com distance_key/value_s; alem de
  longest/max_elev/biggest_territory, grava os PRs de distancia
  (1k/5k/10k/15k/21k/42k, menor tempo) e 1h (maior distancia) a partir
  de run_tracks.best_efforts_jsonb.
- artefatos atualizados e republicados: banco-e-integracoes v1.2
  (runs completo, run_tracks +splits/best_efforts, 2FA no modelo:
  users.totp_activated_at, auth_sessions.mfa_verified, tabela
  mfa_recovery_codes), plano v0.15 (mesmas linhas + intro do jogo
  corrigida para o pivo de territorio permanente).

Testes: bestEfforts (1k/5k/sem 10k numa corrida curta; 1h em metros).
Migracao 00008 reversivel (down testado). E2e contra Postgres real:
ingest com sensores -> metrics de runs + splits -> rollup grava
1k/5k/longest/max_elev.

+++ backend/internal/run/dto.go
@@ -30,8 +30,8 @@ type IngestInput struct {
 type MetricsView struct {
 	RunID         string    `json:"run_id"`
 	Splits        []Split   `json:"splits"`
-	ElevGainM     float64   `json:"elev_gain_m"`
-	ElevLossM     float64   `json:"elev_loss_m"`
+	ElevGainM     int       `json:"elev_gain_m"`
+	ElevLossM     int       `json:"elev_loss_m"`
 	AltMinM       *float64  `json:"alt_min_m,omitempty"`
 	AltMaxM       *float64  `json:"alt_max_m,omitempty"`
 	AvgCadenceSPM *int      `json:"avg_cadence_spm,omitempty"`
